requirements.txt
Django==6.0.1
djangorestframework==3.15.2
mysqlclient==2.2.7
requests==2.32.3
Pillow==11.1.0
tweepy==4.14.0
